requirements.txt
tensorflow==2.10.0
numpy==1.24.0
opencv-python>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.2
scikit-learn>=1.0.2